python/requirements.txt
pandas==2.0.3
scikit-learn==1.3.0
joblib==1.3.2
cloudinary==1.36.0
python-dotenv==1.0.0
validators==0.22.0
